% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{coef.zibr_saem}
\alias{coef.zibr_saem}
\title{Estimated coefficients of a ZIBR fit}
\usage{
\method{coef}{zibr_saem}(object, ...)
}
\arguments{
\item{object}{A \code{zibr_saem} object, the result of \code{\link[=fit_zibr]{fit_zibr()}}.}

\item{...}{Unused, for compatibility with the \code{\link[stats:coef]{stats::coef()}} generic.}
}
\value{
Numeric vector \code{mu} with the coefficients of the logistic part
followed by those of the beta part.
}
\description{
Estimated coefficients of a ZIBR fit
}
