% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{logLik.zibr_saem}
\alias{logLik.zibr_saem}
\title{Marginal log-likelihood of a ZIBR fit}
\usage{
\method{logLik}{zibr_saem}(object, ...)
}
\arguments{
\item{object}{A \code{zibr_saem} object, the result of \code{\link[=fit_zibr]{fit_zibr()}}.}

\item{...}{Unused, for compatibility with the \code{\link[stats:logLik]{stats::logLik()}} generic.}
}
\value{
A \code{logLik} object with the marginal log-likelihood estimated by
importance sampling, with attributes \code{df} (degrees of freedom) and \code{nobs}.
}
\description{
Marginal log-likelihood of a ZIBR fit
}
