% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibbmr.R
\name{print.zibbmr_saem}
\alias{print.zibbmr_saem}
\title{Print a ZIBBMR fit}
\usage{
\method{print}{zibbmr_saem}(x, ...)
}
\arguments{
\item{x}{A \code{zibbmr_saem} object, the result of \code{\link[=fit_zibbmr]{fit_zibbmr()}}.}

\item{...}{Unused, for compatibility with the \code{\link[=print]{print()}} generic.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a ZIBBMR fit
}
