% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saemMicrobiome-package.R
\docType{package}
\name{saemMicrobiome-package}
\alias{saemMicrobiome}
\alias{saemMicrobiome-package}
\title{saemMicrobiome: modelos SAEM para microbioma longitudinal}
\description{
Herramientas para ajustar, simular y comparar modelos mixtos con
inflacion de ceros para datos longitudinales de microbioma:
}
\details{
\itemize{
\item \strong{ZIBR} (zero-inflated beta regression), para proporciones o
abundancias relativas: ver \code{\link[=fit_zibr]{fit_zibr()}}.
\item \strong{ZIBBMR} (zero-inflated beta-binomial mixed regression), para
conteos con profundidad de secuenciacion conocida: ver \code{\link[=fit_zibbmr]{fit_zibbmr()}}.
}

Ambos se estiman con el algoritmo Stochastic Approximation EM (SAEM),
siguiendo la metodologia desarrollada por John Barrera.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabrielagutierrezbernal/saemMicrobiome}
  \item Report bugs at \url{https://github.com/gabrielagutierrezbernal/saemMicrobiome/issues}
}

}
\author{
\strong{Maintainer}: Gabriela Gutierrez \email{gabriella.gutierrez.bernal@gmail.com}

Authors:
\itemize{
  \item Gabriela Gutierrez \email{gabriella.gutierrez.bernal@gmail.com}
}

Other contributors:
\itemize{
  \item John Barrera (Autor original del metodo de estimacion SAEM para ZIBR y ZIBBMR) [contributor]
}

}
\keyword{internal}
