% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibbmr.R
\name{vcov.zibbmr_saem}
\alias{vcov.zibbmr_saem}
\title{Variance-covariance matrix of a ZIBBMR fit}
\usage{
\method{vcov}{zibbmr_saem}(object, ...)
}
\arguments{
\item{object}{A \code{zibbmr_saem} object fitted with \code{compute_fim = TRUE}.}

\item{...}{Unused, for compatibility with the \code{\link[stats:vcov]{stats::vcov()}} generic.}
}
\value{
A variance-covariance matrix.
}
\description{
Variance-covariance matrix of a ZIBBMR fit
}
