% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saemMicrobiome-package.R
\docType{package}
\name{saemMicrobiome-package}
\alias{saemMicrobiome}
\alias{saemMicrobiome-package}
\title{saemMicrobiome: SAEM models for longitudinal microbiome data}
\description{
Tools to fit, simulate and compare zero-inflated mixed models for
longitudinal microbiome data:
}
\details{
\itemize{
\item \strong{ZIBR} (zero-inflated beta regression), for proportions or relative
abundances: see \code{\link[=fit_zibr]{fit_zibr()}}.
\item \strong{ZIBBMR} (zero-inflated beta-binomial mixed regression), for counts
with known sequencing depth: see \code{\link[=fit_zibbmr]{fit_zibbmr()}}.
}

Both are estimated with the Stochastic Approximation EM (SAEM) algorithm,
following the methodology developed by John Barrera.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabrielagutierrezbernal/saemMicrobiome}
  \item Report bugs at \url{https://github.com/gabrielagutierrezbernal/saemMicrobiome/issues}
}

}
\author{
\strong{Maintainer}: Gabriela Gutierrez \email{gabriella.gutierrez.bernal@gmail.com}

Authors:
\itemize{
  \item Gabriela Gutierrez \email{gabriella.gutierrez.bernal@gmail.com}
}

Other contributors:
\itemize{
  \item John Barrera (Original author of the SAEM estimation method for ZIBR and ZIBBMR) [contributor]
}

}
\keyword{internal}
